refactor(eval): use one agentic session auditor
