--- Sources/JarvisCore/Diagnostics/SessionEvaluator.swift
@@ -1,352 +1,0 @@
-import Foundation
-
-/// One-click session evaluation: turns a session's recorded brain traffic (`brain-traffic.jsonl`,
-/// see `BrainTrafficLog`) into a readable wire transcript, appends fixed stop/degrade outcomes from
-/// the human-facing Activity log, sends both to an LLM with an audit prompt focused on context
-/// engineering and user impact, and saves the returned report as `eval-report.md` in the session
-/// directory. This automates the old manual loop of pulling request logs from the OpenAI dashboard
-/// and pasting them into a chat for diagnosis.
-///
-/// The transcript rendering is deliberately delta-aware: the harness rebuilds every request as
-/// `[system] + history + turn` with an append-only prefix, so consecutive requests repeat almost all
-/// of their `input`. Re-sending those repeats to the evaluator would bury the signal (and cost real
-/// tokens), so unchanged instructions/tools/input prefixes are elided and *marked as elided* — the
-/// elision markers themselves tell the evaluator the prompt cache should be hitting there.
-public struct SessionEvaluator: Sendable {
-    public static let reportFilename = "eval-report.md"
-
-    private let brain: BrainClient
-    private let activityLog: ActivityLog?
-
-    public init(brain: BrainClient, activityLog: ActivityLog? = nil) {
-        self.brain = brain
-        self.activityLog = activityLog
-    }
-
-    public enum EvaluationError: LocalizedError, Equatable {
-        case noTraffic
-        case emptyReport
-        public var errorDescription: String? {
-            switch self {
-            case .noTraffic:
-                return "No brain traffic was recorded for this session — nothing to evaluate."
-            case .emptyReport:
-                return "The evaluation model returned an empty report."
-            }
-        }
-    }
-
-    /// The report persisted by an earlier `evaluate`, if any. Lets the viewer reopen a finished
-    /// audit ("Open report") instead of re-billing an evaluation of the same session.
-    public static func savedReport(in sessionDir: URL) -> String? {
-        let url = sessionDir.appendingPathComponent(reportFilename)
-        guard let report = try? String(contentsOf: url, encoding: .utf8), !report.isEmpty
-        else { return nil }
-        return report
-    }
-
-    /// Whether a session has anything to evaluate (a non-empty traffic file).
-    public static func hasTraffic(in sessionDir: URL) -> Bool {
-        let url = sessionDir.appendingPathComponent(BrainTrafficLog.filename)
-        let size = (try? FileManager.default.attributesOfItem(atPath: url.path)[.size] as? NSNumber)
-        return (size?.intValue ?? 0) > 0
-    }
-
-    /// Evaluate one session: render its traffic, ask the brain for the audit, persist and return the
-    /// markdown report. The report is written owner-only next to the traffic it audits, so it survives
-    /// for later browsing and re-reading without re-billing an evaluation.
-    public func evaluate(sessionDir: URL) async throws -> String {
-        // `record` is intentionally asynchronous. Drain it before reading so an Evaluate click
-        // immediately after teardown cannot omit the failure that caused teardown.
-        activityLog?.flush()
-        let url = sessionDir.appendingPathComponent(BrainTrafficLog.filename)
-        let jsonl = (try? String(contentsOf: url, encoding: .utf8)) ?? ""
-        var transcript = Self.renderTranscript(jsonl: jsonl)
-        guard !transcript.isEmpty else { throw EvaluationError.noTraffic }
-        let activityURL = sessionDir.appendingPathComponent(ActivityLog.filename)
-        let activityJSONL = (try? String(contentsOf: activityURL, encoding: .utf8)) ?? ""
-        let activityOutcome = Self.renderActivityOutcome(jsonl: activityJSONL)
-        if !activityOutcome.isEmpty {
-            transcript += "\n\n" + activityOutcome
-        }
-
-        let response = try await brain.respond(
-            messages: [.system(Self.evalInstructions), .user(transcript)],
-            tools: [], toolChoice: .auto)
-        guard let text = response.outputText, !text.isEmpty else {
-            throw EvaluationError.emptyReport
-        }
-        // Stamp provenance so a reader knows this is the wire + Activity-outcome audit, not the
-        // code-reading one.
-        let report = Self.provenanceStamp + "\n\n" + text
-        // A failed write is tolerable: the report is still shown; only "Open report" reuse is lost.
-        _ = FileManager.default.createFile(
-            atPath: sessionDir.appendingPathComponent(Self.reportFilename).path,
-            contents: Data(report.utf8), attributes: [.posixPermissions: 0o600])
-        return report
-    }
-
-    /// Prepended to every single-call report. This path audits the wire transcript plus sanitized
-    /// Activity outcomes; the code-reading agentic audit (`AgenticEvaluation` via
-    /// `scripts/eval-session.sh`) is preferred for a real audit — see the note it stamps.
-    static let provenanceStamp =
-        "> _Produced by the single-call `SessionEvaluator` (in-app wire + Activity-outcome audit). For a "
-        + "code-verified audit with `[confirmed]`/`[hypothesis]` labels, run `scripts/eval-session.sh`._"
-
-    // MARK: - Transcript rendering (pure, testable)
-
-    /// Render the recorded traffic as a readable transcript, one block per round trip, eliding
-    /// request content that is byte-identical to the previous call with the same tag (see the type
-    /// comment). Malformed lines are skipped; an empty/blank file renders as "".
-    static func renderTranscript(jsonl: String) -> String {
-        var blocks: [String] = []
-        // Elision state, per tag (coach and summarizer interleave but never share a prefix).
-        var prevInstructions: [String: String] = [:]
-        var prevTools: [String: String] = [:]
-        var prevInput: [String: [String]] = [:]
-        var callNumber = 0
-
-        for raw in jsonl.split(separator: "\n", omittingEmptySubsequences: true) {
-            guard let entry = (try? JSONSerialization.jsonObject(with: Data(raw.utf8))) as? [String: Any]
-            else { continue }
-            callNumber += 1
-            let tag = entry["tag"] as? String ?? "?"
-            var lines: [String] = []
-
-            var header = "=== call #\(callNumber) · \(tag) · \(entry["t"] as? String ?? "?")"
-            if let status = entry["status"] as? Int { header += " · HTTP \(status)" }
-            if let ms = entry["ms"] as? Int { header += " · \(ms) ms" }
-            lines.append(header)
-            if let error = entry["error"] as? String { lines.append("TRANSPORT ERROR: \(error)") }
-
-            if let request = entry["request"] as? [String: Any] {
-                lines.append(contentsOf: renderRequest(request, tag: tag,
-                                                       prevInstructions: &prevInstructions,
-                                                       prevTools: &prevTools,
-                                                       prevInput: &prevInput))
-            }
-            if let response = entry["response"] {
-                lines.append(contentsOf: renderResponse(response))
-            }
-            blocks.append(lines.joined(separator: "\n"))
-        }
-        guard !blocks.isEmpty else { return "" }
-        // Lead with the deterministic metrics table (computed, not eyeballed) so the auditor
-        // interprets numbers instead of summing usage blobs by hand — the source of the 2026-07-19
-        // audit's cost/cache arithmetic errors. Empty traffic still renders "" (callers guard on it).
-        let body = blocks.joined(separator: "\n\n")
-        return SessionMetrics.render(jsonl: jsonl) + "\n\n" + body
-    }
-
-    /// Render only Activity's typed runtime failure/degrade outcomes. Heard speech, tips,
-    /// screenshots, and raw diagnostics are deliberately excluded: the wire transcript already
-    /// carries the model context, while this compact block answers the missing UX question—did one
-    /// failed call degrade or terminate the user's whole session? Old sessions without `k` retain a
-    /// marker fallback; new sessions never depend on mutable human copy for lifecycle semantics.
-    static func renderActivityOutcome(jsonl: String) -> String {
-        var lines: [String] = []
-        for raw in jsonl.split(separator: "\n", omittingEmptySubsequences: true) {
-            guard let entry = (try? JSONSerialization.jsonObject(with: Data(raw.utf8)))
-                    as? [String: Any],
-                  let message = entry["m"] as? String else { continue }
-            if let rawKind = entry["k"] as? String,
-               let kind = ActivityLog.EventKind(rawValue: rawKind) {
-                guard kind.isEvaluationOutcome else { continue }
-            } else {
-                // Backward compatibility for sessions written before stable event kinds.
-                guard message.hasPrefix("⚠️") || message.hasPrefix("⏹") else { continue }
-            }
-            lines.append("[\(entry["t"] as? String ?? "?")] \(message)")
-        }
-        guard !lines.isEmpty else { return "" }
-        return (["=== human-facing runtime outcome ==="] + lines).joined(separator: "\n")
-    }
-
-    private static func renderRequest(_ request: [String: Any], tag: String,
-                                      prevInstructions: inout [String: String],
-                                      prevTools: inout [String: String],
-                                      prevInput: inout [String: [String]]) -> [String] {
-        var lines: [String] = []
-        var params = "request: model=\(request["model"] as? String ?? "?")"
-        if let reasoning = request["reasoning"] { params += " reasoning=\(compact(reasoning))" }
-        if let cap = request["max_output_tokens"] { params += " max_output_tokens=\(cap)" }
-        if let choice = request["tool_choice"] { params += " tool_choice=\(compact(choice))" }
-        lines.append(params)
-
-        let instructions = request["instructions"] as? String ?? ""
-        if instructions == prevInstructions[tag] {
-            lines.append("instructions: (unchanged — \(instructions.count) chars)")
-        } else if !instructions.isEmpty {
-            lines.append("instructions (\(instructions.count) chars):\n\(instructions)")
-        }
-        prevInstructions[tag] = instructions
-
-        let tools = canonical(request["tools"] ?? [])
-        if tools == prevTools[tag] {
-            let count = (request["tools"] as? [Any])?.count ?? 0
-            lines.append("tools: (unchanged — \(count) defs)")
-        } else {
-            lines.append("tools: \(tools)")
-        }
-        prevTools[tag] = tools
-
-        let items = request["input"] as? [Any] ?? []
-        let rendered = items.map { renderInputItem($0) }
-        let canon = items.map { canonical($0) }
-        let prev = prevInput[tag] ?? []
-        var shared = 0
-        while shared < min(canon.count, prev.count), canon[shared] == prev[shared] { shared += 1 }
-        lines.append("input (\(items.count) items):")
-        if shared > 0 {
-            lines.append("  [items 1–\(shared) unchanged from the previous \(tag) call — the stable, cacheable prefix]")
-        }
-        lines.append(contentsOf: rendered.dropFirst(shared).map { "  \($0)" })
-        prevInput[tag] = canon
-        return lines
-    }
-
-    /// Flatten one Responses `input` item to a single labelled line: role messages get their text
-    /// (image parts were already redacted at record time), function calls/results get name + payload.
-    private static func renderInputItem(_ item: Any) -> String {
-        guard let dict = item as? [String: Any] else { return compact(item) }
-        if let role = dict["role"] as? String {
-            let content = (dict["content"] as? [[String: Any]] ?? [])
-                .compactMap { ($0["text"] ?? $0["image_url"]) as? String }
-                .joined(separator: "\n")
-            return "\(role): \(content)"
-        }
-        switch dict["type"] as? String {
-        case "function_call":
-            return "assistant → function_call \(dict["name"] as? String ?? "?")(\(dict["arguments"] as? String ?? ""))"
-        case "function_call_output":
-            return "tool result: \(dict["output"] as? String ?? "")"
-        case "reasoning":
-            // The tool loop replays reasoning items verbatim (opaque ids, possibly a large
-            // `encrypted_content` blob) — no audit signal in the bytes, so stub them like images.
-            return "assistant reasoning (replayed verbatim — \(compact(dict).count) chars)"
-        // CLI-provider records (`CLIBrainClient`): plain content blocks, images already stubbed.
-        case "text":
-            return "text: \(dict["text"] as? String ?? "")"
-        case "image":
-            return "image: \(dict["image"] as? String ?? "(redacted)")"
-        default:
-            return compact(dict)
-        }
-    }
-
-    private static func renderResponse(_ response: Any) -> [String] {
-        guard let dict = response as? [String: Any] else { return ["response (unparsed): \(compact(response))"] }
-        var lines: [String] = []
-        var header = "response:"
-        if let status = dict["status"] as? String { header += " status=\(status)" }
-        if let details = dict["incomplete_details"] { header += " incomplete_details=\(compact(details))" }
-        if let exit = dict["exitCode"] as? Int { header += " exit=\(exit)" }   // CLI-provider record
-        lines.append(header)
-        for item in dict["output"] as? [[String: Any]] ?? [] {
-            switch item["type"] as? String {
-            case "function_call":
-                lines.append("  → function_call \(item["name"] as? String ?? "?")(\(item["arguments"] as? String ?? ""))")
-            case "message":
-                let text = (item["content"] as? [[String: Any]] ?? [])
-                    .compactMap { $0["text"] as? String }
-                    .joined()
-                lines.append("  → text: \(text)")
-            case "reasoning":
-                continue   // encrypted/empty reasoning stubs carry no signal
-            default:
-                lines.append("  → \(compact(item))")
-            }
-        }
-        // CLI-provider records carry the reply as one string plus the CLI's own envelope metadata
-        // (usage/duration for claude) instead of a Responses `output` array.
-        if let reply = dict["reply"] as? String, !reply.isEmpty { lines.append("  → text: \(reply)") }
-        if let cli = dict["cli"] { lines.append("  cli: \(compact(cli))") }
-        if let stderr = dict["stderr"] as? String, !stderr.isEmpty { lines.append("  stderr: \(stderr)") }
-        // Usage is the quantitative core of the audit: `input_tokens_details.cached_tokens` vs
-        // `input_tokens` is the prompt-cache hit rate, per call.
-        if let usage = dict["usage"] { lines.append("  usage: \(compact(usage))") }
-        if let error = dict["error"], !(error is NSNull) { lines.append("  API ERROR: \(compact(error))") }
-        return lines
-    }
-
-    /// Deterministic single-line JSON for both display and prefix comparison.
-    private static func canonical(_ value: Any) -> String {
-        guard let data = try? JSONSerialization.data(withJSONObject: value,
-                                                     options: [.sortedKeys, .fragmentsAllowed])
-        else { return String(describing: value) }
-        return String(data: data, encoding: .utf8) ?? ""
-    }
-
-    private static func compact(_ value: Any) -> String { canonical(value) }
-
-    // MARK: - The audit prompt
-
-    static let evalInstructions = """
-    You are auditing one session of Jarvis, a proactive macOS coaching assistant: it transcribes \
-    the user's (and their interlocutor's) speech, optionally screenshots the user's screen, and an \
-    LLM decides each turn to speak a short overlay tip, look at the screen, or stay silent. The \
-    session's memory is client-managed: each request is [instructions] + history + new turn, built \
-    append-only so the provider's prompt cache keeps hitting; old history is periodically compacted \
-    into a summary by a separate "summarizer" client.
-
-    The user message opens with a "=== deterministic metrics ===" table computed from the raw traffic \
-    (per-call and aggregate input / cache-read / cache-write / output tokens and cost, plus a \
-    per-model breakdown). Trust its known numbers and availability labels: `—` means unavailable, not \
-    zero, and a `known (N unavailable)` value is partial, not a session total. Quote only what the \
-    table supports; NEVER recompute a total by eye. It follows with the session's complete wire-level \
-    LLM traffic, one block per API \
-    call, in order. To keep it compact: content byte-identical to the previous call with the same tag \
-    is elided and explicitly marked "(unchanged)" — those markers are where the prompt cache SHOULD \
-    be hitting; base64 screenshots are redacted to a stub. When the session recorded a fixed Activity \
-    stop or degrade notice, the user message ends with a "=== human-facing runtime outcome ===" \
-    block. Treat that block as authoritative for user impact: `coaching stopped` / `session failed` \
-    is a session-level UX failure, while `listening continues` is a recoverable missed turn. In \
-    particular, flag a single transport timeout that stopped the whole session as catastrophic and \
-    correlate it with the nearest call error.
-
-    Provider records can appear together — read the right fields for each:
-      - OpenAI Responses: `response.usage` with `input_tokens`, `input_tokens_details.cached_tokens` \
-    (the automatic prefix-cache hit; hit rate = cached_tokens / input_tokens), optional \
-    `cache_write_tokens`, and `output_tokens`; a truncated run shows `status:"incomplete"`. There is \
-    no per-call dollar cost. OpenAI `input_tokens` includes cached input, so mixed-provider token \
-    totals are kept separate.
-      - Local CLI (`claude -p`): `response.cli` with `total_cost_usd`, a call-level `usage` carrying \
-    Anthropic's `cache_creation_input_tokens` / `cache_read_input_tokens` split, and a `modelUsage` \
-    map with per-model usage + cost (including the CLI's internal sidecar models, e.g. a haiku pass). \
-    Anthropic caching is BLOCK-level, not automatic-prefix: a fresh, non-persisted `claude -p` turn \
-    sends the whole conversation as one giant block, so it can only ever cache-hit the reused \
-    `--system-prompt` — a small, flat cache-read across turns is that serialization, NOT the harness \
-    rewriting history. Do not attribute it to cache-busting without checking which envelope this is.
-      - Codex CLI: the recorded response currently has the reply and exit status but no token, cache, \
-    or cost usage. Those cells must stay unavailable; never interpret them as zero.
-
-    Before finalizing, re-check every number you wrote against the metrics table and correct or delete \
-    any that disagree. This transcript ELIDES byte-identical repeats, so it cannot support cardinal \
-    counts of repeated content (stub occurrences, OCR dumps, marker repetitions); take call counts \
-    from the metrics table and, for any other count the table doesn't give, say it is not determinable \
-    from the elided transcript rather than guessing.
-
-    Write a markdown report with exactly these sections:
-
-    ## Context engineering
-    Inefficiencies in what the harness sends: cache-busting prefix changes (input marked as changed \
-    that should have been stable), redundant or stale content re-billed every call, oversized \
-    instructions or tool schemas relative to their value, history-compaction timing/quality, \
-    screenshots kept or stubbed at the wrong time. Be quantitative with the usage numbers.
-
-    ## Issues and errors
-    Transport errors, non-2xx responses, truncated runs (status=incomplete), tool-loop anomalies \
-    (repeated capture_screen, exhausted loops), any contradiction between the instructions and the \
-    model's behavior, and the user-visible lifecycle consequence from the runtime-outcome block.
-
-    ## Coaching quality
-    Given the transcript deltas the model saw: tips that were wrong, late, redundant, or noisy; \
-    silences that should have been tips and vice versa.
-
-    ## Recommendations
-    Concrete changes, ordered by expected impact, each tied to evidence above.
-
-    Cite call numbers (call #N) as evidence throughout. If the data doesn't support a finding, say \
-    so rather than speculating.
-    """
-}
