+++ wiki/settings-window.md
@@ -52,7 +52,7 @@ so an unwrapped fixed-frame view would ride the bottom edge in a taller window).
 | `BrainSection` | "Brain" | yes | Everything that decides who answers a coaching turn, in one tab: the provider radios (OpenAI API / Claude Code / Codex CLI — see [Brain](#brain)), a per-provider model dropdown, the reasoning-effort dropdown (one global setting, mapped onto each provider's scale), and the OpenAI API-key controls (`APIKeyControls`: an `NSSecureTextField` that saves to an owner-only file). Saving a key never restarts a live conversation: established Realtime sockets stay connected and use it on a later reconnect, while an OpenAI brain update uses the same transactional between-turn fallback as other Brain changes. Brain choices take effect on the next coaching turn while running, or the next Start while stopped. |
 | `OverlaySection` | "Overlay" | yes | Two groups, one per overlay surface — **Overlay Caption** (the transient on-screen tip) and **Overlay Box** (the persistent response history). Each has a header with an On/Off toggle (an `NSSwitch` + "On"/"Off" label) and a one-line description. When a surface is **on** it also shows its Text Size + Opacity sliders (with live readouts) and a live sample, **only while the Overlay tab is selected** (`didBecomeActive`/`didResignActive`); when **off**, its sliders and sample are hidden and the layout collapses. Persists via `OverlayAppearance`. |
 | `DisplaySection` | "Screen" | yes | One dropdown — the capture scope: **Active window** (default) or one **Entire display** entry per connected display; persists via `ScreenCapturePreferences`. Applies to the next screenshot. |
-| `ActivitySection` | "Activity" | yes | Embeds the `ActivityViewer` content view (`makeContentView()` / `teardown()`); `fillsTab == true` so the log stretches with the window. Its header shows the selected session's exact directory ID in a selectable field with **Copy ID**, and carries **Evaluate** — one click sends the selected session's recorded LLM wire traffic (`brain-traffic.jsonl`) plus the typed failure/degrade outcomes from `jarvis-activity.jsonl` to the brain model at high effort for a context-engineering and user-impact audit (`SessionEvaluator`), saved as `eval-report.md` in the session dir and opened in the browser as a rendered `eval-report.html` page (`EvalReportPage`) whose **Copy as Markdown** button hands the raw report to an agent chat; once a session has a saved report the button flips to **Open report**. Only *finished* conversations qualify: the button is disabled while the selected session is the live, still-running one (a mid-session audit would judge half a story) and re-enables once Stop has drained any in-flight turn — the cancelled request's final traffic line must land before the audit reads the file. |
+| `ActivitySection` | "Activity" | yes | Embeds the `ActivityViewer` content view (`makeContentView()` / `teardown()`); `fillsTab == true` so the log stretches with the window. Its header shows the selected session's exact directory ID with **Copy ID** and carries **Open report**, enabled only when the dev-side agentic evaluator has written `eval-report.md`. The sole evaluation path is `scripts/eval-session.sh`: its Claude Code / Codex agent receives the complete session directory, reads the full unfiltered `jarvis-activity.jsonl` whenever it needs the user-visible sequence, and correlates it with `brain-traffic.jsonl`, screenshots, and the live source checkout. `EvalReportPage` renders the saved markdown as `eval-report.html`; **Copy as Markdown** hands the raw report to an agent chat. Report opening stays disabled through the live coaching/teardown lifecycle because it presents a browser. |
 
 `AppDelegate` builds the section list at launch and passes it to `SettingsWindow`. All four tabs are
 always present, but each tab's content is created lazily on first selection during that open.
